Move metric core classes to libcore
